FOIA-260: Update component filter response.
Remove data from a jsonapi component data request based on
agency_component ids that are initially included, but do not match the
requested component abbreviation.  This deals with omissions based on
permissions in a more forgiving way, possibly leaving some extra
references in the response to data that was omitted.